+++ src/hyp3_akfire_safe/etc/entrypoint.sh
@@ -1,0 +1,4 @@
+#!/bin/bash --login
+set -e
+conda activate hyp3-akfire-safe
+exec python -um hyp3_akfire_safe "$@"